Fix: 스터디 페이지 비활성화, 별도 컴포넌트 분리, typing 효과 에러 수정
